Refactor model.dart into multiple dart files
